feat: add presets
