+++ src/vitePreview.ts
@@ -1,0 +1,8 @@
+import { Config } from '@playwright/test'
+
+export const vitePreview: Config = {
+  webServer: {
+    command: 'npm run preview',
+    port: 4173
+  }
+}
